Remove unused NonJpaModelBuildItem
I just noticed that `NonJpaModelBuildItem`, which was introduced right
from the beginning of "Hibernate ORM with Panache", is never
instantiated. And from what I can see in the git log, the string `new
NonJpaModelBuildItem` never occurred at any point since the extension
was created.

The (always empty) list of NonJpaModelBuildItem is consumed from many
places in the ORM extension, so removing it would simplify a few things.

And most importantly, it will make the work on XML mapping support easier.

Signed-off-by: Yoann Rodière <yoann@hibernate.org>