Use a the same PU name in the PU descriptors of Reactive PUs as everwhere else
For some reason we were using "default-reactive" instead of "<default>".

This is inconsistent with the constant
PersistentUnitUtil.DEFAULT_PERSISTENCE_UNIT_NAME,
which is used about everwhere, including in CDI qualifiers
(@PersistenceUnit, ...).

This is also inconsistent with conventions in the extension
configuration (where we expect the default PU to be named "<default>"),
and that means any code that attempts to retrieve the runtime configuration
by PU name will not work, since it will mistake the default reactive PU
for a named PU. This is particularly problematic for an upcoming PR of mine.

Hibernate ORM is not supposed to be usable in the same application as
Hibernate Reactive at the moment, so I think we're safe from naming
conflicts.

This will, however, have the side effect of renaming the Hibernate
Reactive PU to "<default>"
in exposed metrics and health checks, as well as in the Dev UI.